create sidekiq job
